ntorus: inital commit
ntorus is an n-dimentional scrolling array.

The array (2d in this case) exists as a section of a larger space; it
has a position and a size. This window can scroll in any direction;
below, the section moves to the right and up. and data that is within
both the old and new sections is preserved, and any data that is shifted
out is reset to a default value.

Callback functions exist for when data leaves the section and enters the
section.

The functions are written to avoid looping over all the data unless
necessicary, as when the window is shifted completly out of the old
position. This mean small shifts are efficent, espically with very large
sections.

^
|
|       +----------+
|  +----|-----+    |
|  |    |ppppp|    |
|  |    |ppppp|    |
|  |    +----------+
|  +----------+
|
+---------------------->